Adding a timer and abbreviations to slur
Adding a timer for validation which shows the time from the start of validation command execution to feedback for each step.

Also doing substitutions so that .x is an abbreviation for constant.x and Fof is an abbreviation for FirstOrderMatching.EFA.  TODO: check if this is broken for quoted strings containing those substrings.